Fixed broken config attributes in NACHA.tdml
Tests were all failing. This is needed because the TDML runner has been
improved to check for invalid config attributes. You can no longer get
away with having a defaultConfig in the testSuite, and then config
attributes where the file is non-existent in the individual test cases.
It will no longer silently ignore the config attributes. If they are
specified, they must be to a file that *is* found, and then it
supercedes the defaultConfig.

The changes here are backward compatible to Daffodil 2.0.0. They're
needed for 2.1.0 development versions and later with the fixed TDML
runner.

Also fixed eclipse classpath so you can see the resources and the scala
code for the tests.