Updated TDML file to specify number of passes for round trip.
TDML runner no longer runs mulitple passes automatically. You must
specify whether tests can do round trips, and exactly how many pasess
are required for them to pass. This eliminates false positives.

The NACHA schemas absorb new-lines, but don't generate them on output,
so except for tests that start from data that contains no newlines, the
tests require two passes so that the unparsed data is cannonicalized to
no-newlines form, which is then re-parsed to get back the expected
infoset.

The test ccd_valid_1 *should* pass with roundTrip="onePass", but the
file has a trailing newline, which is not reproduced, so it requires
twoPass to round trip.

The other tests that take "_nolines.txt" test data files round trip
onePass, as those files have no trailing newline.

This change makes the NACHA TDML files incompatible with Daffodil 2.1.0
unfortunately. The onePass/twoPass feature is on the Daffodil 2.2.0
branch. So this change pushed to a branch named "for-daffodil-2.2.0".