Reorganized files on this branch to follow standard file layout.
This enables many tools to work automatically, including a
regression test framework that tests all portable DFDL schemas against
IBM DFDL and Apache Daffodil.

Unfortunately, this file layout is unexpected for other projects and
tooling. So we have to maintain a branch for this.

Maintaining sbt tweaks to accomodate the original file layout was
too problematic. In theory it is possible. Practically speaking sbt is
too obscure and we were unable to get it to work.

Changes to the NACHA schema on the master branch should be brought over to
this branch to enable automated regression testing.